requirements.txt
Flask==2.2.2
google-cloud-vision==3.1.4
openai==0.28.1
python-dotenv==0.21.0
fpdf2==2.7.7

